+++ src/main/java/edu/ie3/simona/api/mapping/ExtEntityMapping.java
@@ -10,19 +10,11 @@ import edu.ie3.simona.api.simulation.mapping.ExtEntityEntry;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /** Contains the mapping between SIMONA uuid, the external id and the data type the assets hold */
 public class ExtEntityMapping {
 
-  private static final Logger log = LoggerFactory.getLogger(ExtEntityMapping.class);
-
   private final Map<DataType, List<ExtEntityEntry>> extEntities;
-
-  public ExtEntityMapping(Map<DataType, List<ExtEntityEntry>> extEntities) {
-    this.extEntities = extEntities;
-  }
 
   public ExtEntityMapping(List<ExtEntityEntry> extEntityEntryList) {
     this.extEntities =
@@ -45,115 +37,68 @@ public class ExtEntityMapping {
   }
 
   /**
-   * Mapping external id to SIMONA uuid
+   * Returns the full mapping external id to SIMONA uuid. Equals {@code
+   * getExtId2UuidMapping(DataType.values())}.
+   */
+  public Map<String, UUID> getExtId2UuidMapping() {
+    return extEntities.values().stream()
+        .flatMap(Collection::stream)
+        .collect(Collectors.toMap(ExtEntityEntry::id, ExtEntityEntry::uuid));
+  }
+
+  /**
+   * Returns the full mapping SIMONA uuid to external id. Equals {@code
+   * getExtUuid2IdMapping(DataType.values())}.
+   */
+  public Map<UUID, String> getExtUuid2IdMapping() {
+    return extEntities.values().stream()
+        .flatMap(Collection::stream)
+        .collect(Collectors.toMap(ExtEntityEntry::uuid, ExtEntityEntry::id));
+  }
+
+  /**
+   * Mapping external id to SIMONA uuid.
    *
    * @param dataType data type the external asset expects
-   * @return Mapping external id to SIMONA uuid
+   * @return mapping external id to SIMONA uuid
    */
   public Map<String, UUID> getExtId2UuidMapping(DataType dataType) {
     return extEntities.getOrDefault(dataType, Collections.emptyList()).stream()
         .collect(Collectors.toMap(ExtEntityEntry::id, ExtEntityEntry::uuid));
   }
 
-  public Map<String, UUID> getExtId2UuidMapping(DataType... dataType) {
-    return Stream.of(dataType)
+  /**
+   * Mapping external id to SIMONA uuid.
+   *
+   * @param dataTypes the external asset expects
+   * @return mapping external id to SIMONA uuid
+   */
+  public Map<String, UUID> getExtId2UuidMapping(DataType... dataTypes) {
+    return Stream.of(dataTypes)
         .flatMap(type -> extEntities.getOrDefault(type, Collections.emptyList()).stream())
         .collect(Collectors.toMap(ExtEntityEntry::id, ExtEntityEntry::uuid));
   }
 
-  public Map<String, UUID> getFullMapping() {
-    return extEntities.values().stream()
-        .flatMap(Collection::stream)
-        .collect(Collectors.toMap(ExtEntityEntry::id, ExtEntityEntry::uuid));
-  }
-
-  public Map<UUID, String> getFullMappingReverse() {
-    return extEntities.values().stream()
-        .flatMap(Collection::stream)
-        .collect(Collectors.toMap(ExtEntityEntry::uuid, ExtEntityEntry::id));
-  }
-
   /**
-   * Mapping SIMONA uuid to external id
+   * Mapping SIMONA uuid to external id.
    *
    * @param dataType data type the external asset expects
-   * @return Mapping SIMONA uuid to external id
+   * @return mapping SIMONA uuid to external id
    */
   public Map<UUID, String> getExtUuid2IdMapping(DataType dataType) {
     return extEntities.getOrDefault(dataType, Collections.emptyList()).stream()
         .collect(Collectors.toMap(ExtEntityEntry::uuid, ExtEntityEntry::id));
   }
 
-  public Map<UUID, String> getExtUuid2IdMapping(DataType... dataType) {
-    return Stream.of(dataType)
+  /**
+   * Mapping SIMONA uuid to external id.
+   *
+   * @param dataTypes data types the external asset expects
+   * @return mapping SIMONA uuid to external id
+   */
+  public Map<UUID, String> getExtUuid2IdMapping(DataType... dataTypes) {
+    return Stream.of(dataTypes)
         .flatMap(type -> extEntities.getOrDefault(type, Collections.emptyList()).stream())
         .collect(Collectors.toMap(ExtEntityEntry::uuid, ExtEntityEntry::id));
-  }
-
-  public static List<UUID> toSimona(List<String> list, Map<String, UUID> mapping) {
-    return list.stream()
-        .map(str -> Optional.ofNullable(mapping.get(str)))
-        .filter(Optional::isPresent)
-        .map(Optional::get)
-        .toList();
-  }
-
-  public static List<String> toExt(List<UUID> list, Map<UUID, String> mapping) {
-    return list.stream()
-        .map(str -> Optional.ofNullable(mapping.get(str)))
-        .filter(Optional::isPresent)
-        .map(Optional::get)
-        .toList();
-  }
-
-  /**
-   * Maps a map: ids to value to a map: uuids to value.
-   *
-   * @param valueMap that should be remapped
-   * @param mapping that is used
-   * @return a map: uuids to value
-   * @param <T> type of output values
-   * @param <V> type of input values
-   */
-  @SuppressWarnings("unchecked")
-  public static <T, V> Map<UUID, T> mapToSimona(
-      Map<String, V> valueMap, Map<String, UUID> mapping) {
-    Map<UUID, T> convertedMap = new HashMap<>();
-
-    for (String key : valueMap.keySet()) {
-      V value = valueMap.get(key);
-
-      try {
-        convertedMap.put(mapping.get(key), (T) value);
-      } catch (ClassCastException e) {
-        log.warn("Could not convert value {} for key {}", value, key, e);
-      }
-    }
-    return convertedMap;
-  }
-
-  /**
-   * Maps a map: uuids to value to a map: ids to value.
-   *
-   * @param valueMap that should be remapped
-   * @param mapping that is used
-   * @return a map: id to value
-   * @param <T> type of output values
-   * @param <V> type of input values
-   */
-  @SuppressWarnings("unchecked")
-  public static <T, V> Map<String, T> mapToExt(Map<UUID, V> valueMap, Map<UUID, String> mapping) {
-    Map<String, T> convertedMap = new HashMap<>();
-
-    for (UUID key : valueMap.keySet()) {
-      V value = valueMap.get(key);
-
-      try {
-        convertedMap.put(mapping.get(key), (T) value);
-      } catch (ClassCastException e) {
-        log.warn("Could not convert value {} for key {}", value, key, e);
-      }
-    }
-    return convertedMap;
   }
 }