+++ src/main/java/edu/ie3/simona/api/ontology/em/EmCompletion.java
@@ -8,4 +8,9 @@ package edu.ie3.simona.api.ontology.em;
 
 import java.util.Optional;
 
+/**
+ * Response send from SIMONA after the em service is finished.
+ *
+ * @param maybeNextTick option for the next tick in SIMONA
+ */
 public record EmCompletion(Optional<Long> maybeNextTick) implements EmDataResponseMessageToExt {}