+++ src/main/java/edu/ie3/simona/api/ontology/em/RequestEmCompletion.java
@@ -6,4 +6,9 @@
 
 package edu.ie3.simona.api.ontology.em;
 
+/**
+ * Request send to SIMONA to finish the em service.
+ *
+ * @param tick for which the em service should be finished
+ */
 public record RequestEmCompletion(long tick) implements EmDataMessageFromExt {}